fix: make modals work again
